Add memory management
Relying on golang to recycle memory causes memory not to be returned
to the operating system. Instead, keep our own pool of buffers and
use those when available, shrinking and returning to the operating
system to the extent unused for 5 seconds. The pool of buffers can
grow to twice the size required for the largest request.

Signed-off-by: Alex Bligh <alex@alex.org.uk>